@@ -1,7 +1,7 @@
 {
     "name": "sourcetoad/enhanced-resources",
     "description": "Laravel's API Resources enhanced.",
-    "minimum-stability": "stable",
+    "minimum-stability": "dev",
     "license": "MIT",
     "authors": [
         {
@@ -20,11 +20,11 @@
         }
     },
     "require": {
-        "php": "^8.0",
-        "laravel/framework": "^9.0|^10.0"
+        "php": "^8.2",
+        "laravel/framework": "^10.0|^11.0|11.x-dev"
     },
     "require-dev": {
-        "orchestra/testbench": "^7.0|^8.0",
-        "phpunit/phpunit": "^9.5|^10.0"
+        "orchestra/testbench": "^8.0|^9.0|9.x-dev",
+        "phpunit/phpunit": "^10.0|^11.0"
     }
 }
